refactor: split ProductCollection test file into focused files
