+++ plugins/woocommerce/tests/php/src/Blocks/BlockTypes/ProductCollection/ProductCollectionQueryBuilder.php
@@ -1,0 +1,700 @@
+<?php
+
+namespace Automattic\WooCommerce\Tests\Blocks\BlockTypes\ProductCollection;
+
+use Automattic\WooCommerce\Tests\Blocks\Mocks\ProductCollectionMock;
+use WC_Helper_Product;
+use WP_Query;
+use Automattic\WooCommerce\Enums\ProductStockStatus;
+
+/**
+ * Tests for the ProductCollection block query building and merging logic
+ *
+ * @group query-building
+ */
+class ProductCollectionQueryBuilder extends \WP_UnitTestCase {
+	/**
+	 * This variable holds our Product Query object.
+	 *
+	 * @var ProductCollectionMock
+	 */
+	private $block_instance;
+
+	/**
+	 * Return starting point for parsed block test data.
+	 * Using a method instead of property to avoid sharing data between tests.
+	 */
+	private function get_base_parsed_block() {
+		return array(
+			'blockName' => 'woocommerce/product-collection',
+			'attrs'     => array(
+				'query' => array(
+					'perPage'                  => 9,
+					'pages'                    => 0,
+					'offset'                   => 0,
+					'postType'                 => 'product',
+					'order'                    => 'desc',
+					'orderBy'                  => 'date',
+					'search'                   => '',
+					'exclude'                  => array(),
+					'sticky'                   => '',
+					'inherit'                  => true,
+					'isProductCollectionBlock' => true,
+					'woocommerceAttributes'    => array(),
+					'woocommerceStockStatus'   => array(
+						ProductStockStatus::IN_STOCK,
+						ProductStockStatus::OUT_OF_STOCK,
+						ProductStockStatus::ON_BACKORDER,
+					),
+				),
+			),
+		);
+	}
+
+	/**
+	 * Initiate the mock object.
+	 */
+	protected function setUp(): void {
+		$this->block_instance = new ProductCollectionMock();
+	}
+
+	/**
+	 * Build the merged_query for testing
+	 *
+	 * @param array $parsed_block Parsed block data.
+	 * @param array $query        Query data.
+	 */
+	private function initialize_merged_query( $parsed_block = array(), $query = array() ) {
+		if ( empty( $parsed_block ) ) {
+			$parsed_block = $this->get_base_parsed_block();
+		}
+
+		$this->block_instance->set_parsed_block( $parsed_block );
+
+		$block          = new \stdClass();
+		$block->context = $parsed_block['attrs'];
+
+		return $this->block_instance->build_frontend_query( $query, $block, 1 );
+	}
+
+	/**
+	 * Test merging featured queries.
+	 */
+	public function test_merging_featured_queries() {
+		$parsed_block                               = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['featured'] = true;
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertContainsEquals(
+			array(
+				'field'    => 'name',
+				'terms'    => 'featured',
+				'operator' => 'IN',
+				'taxonomy' => 'product_visibility',
+			),
+			$merged_query['tax_query']
+		);
+	}
+
+	/**
+	 * Test merging on sale queries.
+	 */
+	public function test_merging_on_sale_queries() {
+		// Mock the on sale product ids.
+		$on_sale_product_ids = array( 1, 2, 3, 4 );
+		set_transient( 'wc_products_onsale', $on_sale_product_ids, DAY_IN_SECONDS * 30 );
+
+		$parsed_block                                        = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['woocommerceOnSale'] = true;
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		foreach ( $on_sale_product_ids as $id ) {
+			$this->assertContainsEquals( $id, $merged_query['post__in'] );
+		}
+
+		$this->assertCount( 4, $merged_query['post__in'] );
+
+		delete_transient( 'wc_products_onsale' );
+	}
+
+	/**
+	 * Test merging stock status queries.
+	 */
+	public function test_merging_stock_status_queries() {
+		$parsed_block = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['woocommerceStockStatus'] = array(
+			ProductStockStatus::OUT_OF_STOCK,
+			ProductStockStatus::ON_BACKORDER,
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertContainsEquals(
+			array(
+				'value'   => array( ProductStockStatus::OUT_OF_STOCK, ProductStockStatus::ON_BACKORDER ),
+				'compare' => 'IN',
+				'key'     => '_stock_status',
+			),
+			$merged_query['meta_query']
+		);
+	}
+
+	/**
+	 * Test merging default stock queries that should use product visibility
+	 * queries instead of meta query for stock status.
+	 */
+	public function test_merging_default_stock_queries() {
+		$parsed_block = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['woocommerceStockStatus'] = array(
+			ProductStockStatus::IN_STOCK,
+			ProductStockStatus::OUT_OF_STOCK,
+			ProductStockStatus::ON_BACKORDER,
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEmpty( $merged_query['meta_query'] );
+
+		// Test with hide out of stock items option enabled.
+		$parsed_block = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['woocommerceStockStatus'] = array(
+			ProductStockStatus::IN_STOCK,
+			ProductStockStatus::ON_BACKORDER,
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEmpty( $merged_query['meta_query'] );
+	}
+
+	/**
+	 * Test merging attribute queries.
+	 */
+	public function test_merging_attribute_queries() {
+		$parsed_block = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['woocommerceAttributes'] = array(
+			array(
+				'taxonomy' => 'pa_test',
+				'termId'   => 1,
+			),
+			array(
+				'taxonomy' => 'pa_test',
+				'termId'   => 2,
+			),
+			array(
+				'taxonomy' => 'pa_another_test',
+				'termId'   => 3,
+			),
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertContainsEquals(
+			array(
+				'field'    => 'term_id',
+				'terms'    => array( 3 ),
+				'operator' => 'IN',
+				'taxonomy' => 'pa_another_test',
+			),
+			$merged_query['tax_query']
+		);
+
+		$this->assertContainsEquals(
+			array(
+				'taxonomy' => 'pa_test',
+				'field'    => 'term_id',
+				'terms'    => array( 1, 2 ),
+				'operator' => 'IN',
+			),
+			$merged_query['tax_query']
+		);
+	}
+
+	/**
+	 * Test merging order by rating queries.
+	 */
+	public function test_merging_order_by_rating_queries() {
+		$parsed_block                              = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['orderBy'] = 'rating';
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEquals( 'meta_value_num', $merged_query['orderby'] );
+		$this->assertEquals( '_wc_average_rating', $merged_query['meta_key'] );
+	}
+
+	/**
+	 * Test product visibility query exist in merged query.
+	 */
+	public function test_product_visibility_query_exist_in_merged_query() {
+		$product_visibility_terms  = wc_get_product_visibility_term_ids();
+		$product_visibility_not_in = array( is_search() ? $product_visibility_terms['exclude-from-search'] : $product_visibility_terms['exclude-from-catalog'] );
+
+		$parsed_block = $this->get_base_parsed_block();
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertContainsEquals(
+			array(
+				'taxonomy' => 'product_visibility',
+				'field'    => 'term_taxonomy_id',
+				'terms'    => $product_visibility_not_in,
+				'operator' => 'NOT IN',
+			),
+			$merged_query['tax_query']
+		);
+	}
+
+	/**
+	 * Test merging multiple queries.
+	 */
+	public function test_merging_multiple_queries() {
+		$parsed_block                              = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['orderBy'] = 'rating';
+		$parsed_block['attrs']['query']['woocommerceStockStatus'] = array(
+			ProductStockStatus::IN_STOCK,
+			ProductStockStatus::OUT_OF_STOCK,
+		);
+		$parsed_block['attrs']['query']['woocommerceAttributes']  = array(
+			array(
+				'taxonomy' => 'pa_test',
+				'termId'   => 1,
+			),
+			array(
+				'taxonomy' => 'pa_test',
+				'termId'   => 2,
+			),
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEquals( 'meta_value_num', $merged_query['orderby'] );
+		$this->assertEquals( '_wc_average_rating', $merged_query['meta_key'] );
+		$this->assertContainsEquals(
+			array(
+				'compare' => 'IN',
+				'key'     => '_stock_status',
+				'value'   => array( ProductStockStatus::IN_STOCK, ProductStockStatus::OUT_OF_STOCK ),
+			),
+			$merged_query['meta_query']
+		);
+		$this->assertContainsEquals(
+			array(
+				'taxonomy' => 'pa_test',
+				'field'    => 'term_id',
+				'terms'    => array( 1, 2 ),
+				'operator' => 'IN',
+			),
+			$merged_query['tax_query']
+		);
+	}
+
+	/**
+	 * Test merging filter by max price queries.
+	 */
+	public function test_merging_filter_by_max_price_queries() {
+		set_query_var( 'max_price', 100 );
+
+		$merged_query = $this->initialize_merged_query();
+
+		$this->assertContainsEquals(
+			array(
+				array(
+					'key'     => '_price',
+					'value'   => 100,
+					'compare' => '<=',
+					'type'    => 'numeric',
+				),
+				array(),
+				'relation' => 'AND',
+			),
+			$merged_query['meta_query']
+		);
+		set_query_var( 'max_price', '' );
+	}
+
+	/**
+	 * Test merging filter by min price queries.
+	 */
+	public function test_merging_filter_by_min_price_queries() {
+		set_query_var( 'min_price', 20 );
+
+		$merged_query = $this->initialize_merged_query();
+
+		$this->assertContainsEquals(
+			array(
+				array(),
+				array(
+					'key'     => '_price',
+					'value'   => 20,
+					'compare' => '>=',
+					'type'    => 'numeric',
+				),
+				'relation' => 'AND',
+			),
+			$merged_query['meta_query']
+		);
+		set_query_var( 'min_price', '' );
+	}
+
+	/**
+	 * Test merging filter by min and max price queries.
+	 */
+	public function test_merging_filter_by_min_and_max_price_queries() {
+		set_query_var( 'max_price', 100 );
+		set_query_var( 'min_price', 20 );
+
+		$merged_query = $this->initialize_merged_query();
+
+		$this->assertContainsEquals(
+			array(
+				array(
+					'key'     => '_price',
+					'value'   => 100,
+					'compare' => '<=',
+					'type'    => 'numeric',
+				),
+				array(
+					'key'     => '_price',
+					'value'   => 20,
+					'compare' => '>=',
+					'type'    => 'numeric',
+				),
+				'relation' => 'AND',
+			),
+			$merged_query['meta_query']
+		);
+
+		set_query_var( 'max_price', '' );
+		set_query_var( 'min_price', '' );
+	}
+
+	/**
+	 * Test merging filter by stock status queries.
+	 */
+	public function test_merging_filter_by_stock_status_queries() {
+		set_query_var( 'filter_stock_status', ProductStockStatus::IN_STOCK );
+
+		$merged_query = $this->initialize_merged_query();
+
+		$this->assertContainsEquals(
+			array(
+				'operator' => 'IN',
+				'key'      => '_stock_status',
+				'value'    => array( ProductStockStatus::IN_STOCK ),
+			),
+			$merged_query['meta_query']
+		);
+
+		set_query_var( 'filter_stock_status', '' );
+	}
+
+	/**
+	 * Test merging time range queries.
+	 */
+	public function test_merging_time_frame_before_queries() {
+		$time_frame_date = gmdate( 'Y-m-d H:i:s' );
+
+		$parsed_block                                = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['timeFrame'] = array(
+			'operator' => 'not-in',
+			'value'    => $time_frame_date,
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertContainsEquals(
+			array(
+				'column'    => 'post_date_gmt',
+				'before'    => $time_frame_date,
+				'inclusive' => true,
+			),
+			$merged_query['date_query'],
+		);
+	}
+
+	/**
+	 * Test merging time range queries.
+	 */
+	public function test_merging_time_frame_after_queries() {
+		$time_frame_date = gmdate( 'Y-m-d H:i:s' );
+
+		$parsed_block                                = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['timeFrame'] = array(
+			'operator' => 'in',
+			'value'    => $time_frame_date,
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertContainsEquals(
+			array(
+				'column'    => 'post_date_gmt',
+				'after'     => $time_frame_date,
+				'inclusive' => true,
+			),
+			$merged_query['date_query'],
+		);
+	}
+
+	/**
+	 * Test merging filter by stock status queries.
+	 */
+	public function test_merging_filter_by_attribute_queries() {
+		// Mock the attribute data.
+		$this->block_instance->set_attributes_filter_query_args(
+			array(
+				array(
+					'filter'     => 'filter_color',
+					'query_type' => 'query_type_color',
+				),
+				array(
+					'filter'     => 'filter_size',
+					'query_type' => 'query_type_size',
+				),
+			)
+		);
+
+		set_query_var( 'filter_color', 'blue' );
+		set_query_var( 'query_type_color', 'or' );
+		set_query_var( 'filter_size', 'xl,xxl' );
+		set_query_var( 'query_type_size', 'and' );
+
+		$merged_query = $this->initialize_merged_query();
+		$tax_queries  = $merged_query['tax_query'];
+
+		$and_query = array();
+		foreach ( $tax_queries as $tax_query ) {
+			if ( isset( $tax_query['relation'] ) && 'AND' === $tax_query['relation'] ) {
+				$and_query = $tax_query;
+			}
+		}
+
+		// Check if the AND query is an array.
+		$this->assertIsArray( $and_query );
+
+		$attribute_queries = array();
+		foreach ( $and_query as $and_query_item ) {
+			if ( is_array( $and_query_item ) ) {
+				$attribute_queries = $and_query_item;
+			}
+		}
+
+		$this->assertContainsEquals(
+			array(
+				'taxonomy' => 'pa_color',
+				'field'    => 'slug',
+				'terms'    => array( 'blue' ),
+				'operator' => 'IN',
+			),
+			$attribute_queries
+		);
+
+		$this->assertContainsEquals(
+			array(
+				'taxonomy' => 'pa_size',
+				'field'    => 'slug',
+				'terms'    => array( 'xl', 'xxl' ),
+				'operator' => 'AND',
+			),
+			$attribute_queries
+		);
+
+		set_query_var( 'filter_color', '' );
+		set_query_var( 'query_type_color', '' );
+		set_query_var( 'filter_size', '' );
+		set_query_var( 'query_type_size', '' );
+	}
+
+	/**
+	 * Test merging multiple filter queries.
+	 */
+	public function test_merging_multiple_filter_queries() {
+		set_query_var( 'max_price', 100 );
+		set_query_var( 'min_price', 20 );
+		set_query_var( 'filter_stock_status', ProductStockStatus::IN_STOCK );
+
+		$merged_query = $this->initialize_merged_query();
+
+		$this->assertContainsEquals(
+			array(
+				'operator' => 'IN',
+				'key'      => '_stock_status',
+				'value'    => array( ProductStockStatus::IN_STOCK ),
+			),
+			$merged_query['meta_query']
+		);
+
+		$this->assertContainsEquals(
+			array(
+				array(
+					'key'     => '_price',
+					'value'   => 100,
+					'compare' => '<=',
+					'type'    => 'numeric',
+				),
+				array(
+					'key'     => '_price',
+					'value'   => 20,
+					'compare' => '>=',
+					'type'    => 'numeric',
+				),
+				'relation' => 'AND',
+			),
+			$merged_query['meta_query']
+		);
+
+		set_query_var( 'max_price', '' );
+		set_query_var( 'min_price', '' );
+		set_query_var( 'filter_stock_status', '' );
+	}
+
+	/**
+	 * Test merging taxonomies query i.e.
+	 * - Product categories
+	 * - Product tags
+	 */
+	public function test_merging_taxonomies_query() {
+		$merged_query = $this->initialize_merged_query(
+			null,
+			// Since we aren't calling the Query Loop build function, we need to provide
+			// a tax_query rather than relying on it generating one from the input.
+			array(
+				// phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_tax_query
+				'tax_query' => array(
+					array(
+						'taxonomy'         => 'product_cat',
+						'terms'            => array( 1, 2 ),
+						'include_children' => false,
+					),
+					array(
+						'taxonomy'         => 'product_tag',
+						'terms'            => array( 3, 4 ),
+						'include_children' => false,
+					),
+				),
+			)
+		);
+
+		$this->assertContains(
+			array(
+				'taxonomy'         => 'product_cat',
+				'terms'            => array( 1, 2 ),
+				'include_children' => false,
+			),
+			$merged_query['tax_query']
+		);
+
+		$this->assertContains(
+			array(
+				'taxonomy'         => 'product_tag',
+				'terms'            => array( 3, 4 ),
+				'include_children' => false,
+			),
+			$merged_query['tax_query']
+		);
+	}
+
+	/**
+	 * Test that price range queries are set so they can be picked up in the `posts_clauses` filter.
+	 */
+	public function test_price_range_queries() {
+		$parsed_block                                 = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['priceRange'] = array(
+			'min' => 1,
+			'max' => 100,
+		);
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEquals(
+			array(
+				'min' => 1,
+				'max' => 100,
+			),
+			$merged_query['priceRange'],
+		);
+	}
+
+	/**
+	 * Test handpicked products queries.
+	 */
+	public function test_handpicked_products_queries() {
+		$handpicked_product_ids = array( 1, 2, 3, 4 );
+
+		$parsed_block = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['woocommerceHandPickedProducts'] = $handpicked_product_ids;
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		foreach ( $handpicked_product_ids as $id ) {
+			$this->assertContainsEquals( $id, $merged_query['post__in'] );
+		}
+
+		$this->assertCount( 4, $merged_query['post__in'] );
+	}
+
+	/**
+	 * Test merging exclusive id filters.
+	 */
+	public function test_merges_post__in() {
+		$existing_id_filter     = array( 1, 4 );
+		$handpicked_product_ids = array( 3, 4, 5, 6 );
+		// The only ID present in ALL of the exclusive filters is 4.
+		$expected_product_ids = array( 4 );
+
+		$parsed_block                               = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['post__in'] = $existing_id_filter;
+		$parsed_block['attrs']['query']['woocommerceHandPickedProducts'] = $handpicked_product_ids;
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		foreach ( $expected_product_ids as $id ) {
+			$this->assertContainsEquals( $id, $merged_query['post__in'] );
+		}
+
+		$this->assertCount( 1, $merged_query['post__in'] );
+	}
+
+	/**
+	 * Test merging exclusive id filters with no intersection.
+	 */
+	public function test_merges_post__in_empty_result_without_intersection() {
+		$existing_id_filter     = array( 1, 4 );
+		$handpicked_product_ids = array( 2, 3 );
+
+		$parsed_block                               = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['post__in'] = $existing_id_filter;
+		$parsed_block['attrs']['query']['woocommerceHandPickedProducts'] = $handpicked_product_ids;
+
+		$merged_query = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEquals( array( -1 ), $merged_query['post__in'] );
+	}
+
+	/**
+	 * Test the menu_order sorting functionality.
+	 */
+	public function test_menu_order_sorting() {
+		$parsed_block                              = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['orderBy'] = 'menu_order';
+		$parsed_block['attrs']['query']['order']   = 'asc';
+		$merged_query                              = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEquals( 'menu_order', $merged_query['orderby'] );
+		$this->assertEquals( 'asc', $merged_query['order'] );
+	}
+
+	/**
+	 * Test the random sorting functionality.
+	 */
+	public function test_random_sorting() {
+		$parsed_block                              = $this->get_base_parsed_block();
+		$parsed_block['attrs']['query']['orderBy'] = 'random';
+		$merged_query                              = $this->initialize_merged_query( $parsed_block );
+
+		$this->assertEquals( 'rand', $merged_query['orderby'] );
+	}
+}
